+++ clang/docs/ReleaseNotes.rst
@@ -707,6 +707,7 @@ Improvements to Clang's diagnostics
     }
 
 - Fix -Wdangling false positives on conditional operators (#120206).
+- Clang now diagnoses unused private fields with the ``[[warn_unused]]`` attribute (#GH62472).
 
 Improvements to Clang's time-trace
 ----------------------------------